+++ R/export_dirs.R
@@ -4,7 +4,7 @@
 #'
 #'@param config_file name of the master LakeEnsemblR config file
 #'@param model vector; model to export configuration file.
-#'  Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake")
+#'  Options include c("GOTM", "GLM", "Simstrat", "FLake", "MyLake", "air2water")
 #'@param folder folder
 #'@keywords methods
 #'@examples
@@ -12,7 +12,7 @@
 #'
 #'@export
 
-export_dirs <- function(config_file, model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake"),
+export_dirs <- function(config_file, model = c("GOTM", "GLM", "Simstrat", "FLake", "MyLake", "air2water"),
                           folder = "."){
   # Set working directory
   oldwd <- getwd()
@@ -100,5 +100,32 @@ export_dirs <- function(config_file, model = c("GOTM", "GLM", "Simstrat", "FLake
     }
   }
   
-  message("export_dirs complete!")
+  ##---------------air2water-------------
+  if("air2water" %in% model){
+    lakename <- get_yaml_value(config_file, "location", "name")
+    # Create directory and output directory, if they do not yet exist
+    if(!dir.exists("air2water")){
+      dir.create("air2water")
+    }
+    # create subfolder for forcings
+    if(!dir.exists(file.path("air2water", lakename))){
+      dir.create(file.path("air2water", lakename))
+    }
+    
+    # Load config file
+    temp_fil <- get_yaml_value(config_file, "config_files", "air2water")
+    if(!file.exists(temp_fil)){
+      get_template("air2water_config", folder = folder, filename = temp_fil)
+    }
+    file.copy(file.path(system.file("extdata", package = "air2wateR"), "PSO.txt"),
+              file.path(folder, "air2water"), overwrite = FALSE)
+    # change lake name in the input file
+    a2w_config <- readLines(temp_fil)
+    a2w_config[2] <- gsub("Superior", lakename, a2w_config[2])
+    con <- file(temp_fil, encoding = "ASCII")
+    writeLines(a2w_config, con)
+    close(con)
+  
+    message("export_dirs complete!")
+  }
 }